Webapp/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
pandas==2.2.2

